deploy(playwright): work around externally-enforced HTTPS
It is possible to enforce HTTPS on GitHub Pages, via
https://docs.github.com/en/pages/getting-started-with-github-pages/securing-your-github-pages-site-with-https#enforcing-https-for-your-github-pages-site

Without this flag enabled, it is still possible to access the site via
http:// URLs, otherwise such URLs will automatically redirect to
https:// URLs.

However, it is also possible to enforce HTTPS e.g. by using a custom
domain and putting Cloudflare in front of GitHub Pages.

The Playwright tests, which are designed to catch undesired redirects
that may, say, drop the `/git-scm.com/` part of the URLs in forks by
mistake, are currently not expecting this.

Since GitHub Pages supports HTTPS in all cases, and there are no options
anywhere to enforce HTTP instead of HTTPS, let's just test with the
https:// URL even if no enforcement was configured.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>