ci: do open an issue when a broken link is detected
As of v2.47.0, there is a problem where `/docs/platform-support` links
to `/howto/maintain-git` (it should be prefixed with `/docs` is the
bug).

And indeed, directing a web browser to
https://git-scm.com/docs/platform-support and following the "How to
maintain Git" link will lead to a 404.

This problem was identified correctly by
https://github.com/git/git-scm.com/actions/runs/11219313103, but when
that workflow run tried to open a ticket about the problem, it failed
https://github.com/git/git-scm.com/actions/runs/11219313103/job/31184984303#step:11:127081

	ReferenceError: req is not defined
	    at eval (eval at callAsyncFunction (/home/runner/work/_actions/actions/github-script/v7/dist/index.js:35424:16), <anonymous>:33:42)
	    at process.processTicksAndRejections (node:internal/process/task_queues:95:5)
	    at async main (/home/runner/work/_actions/actions/github-script/v7/dist/index.js:35522:20)
	Error: Unhandled error: ReferenceError: req is not defined

The problem is that the variable `req` was defined within too narrow a
scope, which is fixed by this here commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>