deploy: fix bogus link check for `../`-style links
In 72d3f4f5b (ci(lychee): do detect incorrect `../` links, 2024-10-07),
I introduced a somewhat complicated logic to detect bogus `../`-style
links like the `link:../howto/maintain-git.txt` one introduced via Git
v2.47.0.

The main idea was to do introduce the step output `remap-dotdot` which
would be only set in forks but not in the main repository.

However, due to the buggy implementation, it would be set in the main
repository, to:

	--remap '(http:/.*) file://../$1'

... which basically marked all http:// URLs as broken links.

Let's fix that logic so that `remap-dotdot` _is_ empty in the main
repository.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>