update-download-data: accommodate for Git for Windows' four-part version format
Git for Windows' versions are usually based on upstream Git versions,
and hence uses the same version string.

Every once in a while, an extra version is required, in which case Git
for Windows appends an extra number in parentheses, starting with 2
(implying that regular Git for Windows versions are implicitly "(1)"),
at least where parentheses can be used (otherwise it appends it
separated by a dot).

Accommodate for such versions in the script that wants to update the
download data.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>