Add a script to figure out the ProGit v1 URLs
In 2014, the second edition of the ProGit book was started, in 2016 this
edition became the default, and in 2020 all v1 URLs were redirected to
the landing page of v2.

At some stage, the v1 links stopped working, returning a "500 Internal
server error".

After switching the git-scm.com site to a static Hugo-generated site,
those links now return a "404 That page doesn't exist".

This is far from an ideal situation.

Unfortunately, in a static site, there is no way to install wildcard
routes that would redirect to a better URL, therefore we have to
enumerate all of the URLs that we want to redirect.

Internet Archive to the rescue!

This script downloads the tables of contents of the ProGit v1 book and
its translations, and determines the URLs that had been active back
then, maps them to the v2 equivalents (on a best effort basis), and then
writes out a YAML file with that information.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>