book: add aliases for the old ProGit v1 links
Once upon a time, there first edition of the ProGit book was available
on Git's home page, and the sun was shining. Then, one day in 2014, the
sun shone brighter and work was begun to write the second edition of
this book. At some stage, this became the default when directing web
browsers to https://git-scm.com/book, and a few years later, 2020 or so,
the links that formerly led to the first edition would redirect to v2.

Then, one day, clouds moved across the sky and the redirects from v1 to
v2 stopped working and instead a "500 Internal server error" page was
shown.

Time went by and nobody really knew how to fix it (or more likely,
wasn't in the mood, or wanted other people to fix it).

Finally, in the fall of 2024, git-scm.com was switched to a static web
site, generated using Hugo, and local development became much easier.
Naturally, the v1-to-v2 redirects were no longer in place and the v1
links therefore showed 500 no longer, but a 404.

Still, nobody knew how to fix it, or wasn't in the mood, or wanted other
people to fix it for them.

Until now. Now is the day when we resurrect the v1-to-v2 redirects, in
even more glory than ever before, for now we redirect to the v2 sections
that correspond to the v1 sections (as far as possible, that is)!

Only one (slight) fly in the ointment: URLs to v1 sections of the book
which contain anchors will keep those anchors as-are, and not translate
them to the corresponding new anchors. Example:
Git-Basics-Getting-a-Git-Repository#Cloning-an-Existing-Repository
should redirect to v2/Git-Basics-Getting-a-Git-Repository#_git_cloning,
but does not. It redirects to that page but still tries to find the
anchor `#Cloning-an-Existing-Repository`.

Alas, this is where I do not know how to fix it, or ain't in the mood,
or want other people to fix it for themselves.

This commit addresses https://github.com/git/git-scm.com/issues/1782

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>