ci(update-book): the English version owns `/book/index.html`
The default language of the ProGit book is English, therefore
https://git-scm.com/book should redirect to the table of contents of the
English version of that book.

This means that the `/external/book/content/book/en/_index.html` file
needs to be part of the sparse checkout, otherwise the workflow run
would not be able to update it (should it ever become necessary).

This was not a problem so far because that file remained unchanged (and
is likely to remain so for quite some time yet).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>