Keep sidebar open when clicking inside it
The "sidebar menu" is implemented via two <div>s, one acting as the
button (`.sidebar-btn`), the other as the menu (`.sidebar`).

The menu is shown by virtue of a `.sidebar-btn:focus + .sidebar` rule
that sets the `transform` attribute to 0 (it defaults to -105.5%).

The big problem with this design is that clicking inside the sidebar
menu transfers focus away from the button, and the sidebar menu
vanishes. Typically without following the link that the viewer clicked,
as reported in https://github.com/git/git-scm.com/issues/1916.

Let's let the sidebar menu be shown also when the focus is within said
menu, via the `:focus-within` pseudo-class (for more details, see
https://developer.mozilla.org/en-US/docs/Web/CSS/:focus-within), which
is supported reasonably widely by browsers.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>