deploy: work around the grayed out "Enforce HTTPS" checkbox
We do want to enforce HTTPS. And we kind of do, but not "kind of" enough
for GitHub's UI at https://github.com/git/git-scm.com/settings/pages to
allow us to check the box called "Enforce HTTPS".

The helpful comment next to the grayed out box says:

  [ ] **Enforce HTTPS** --- Unavailable for your site because your
  domain is not properly configured to support HTTPS
  ([`git-scm.com`](http://git-scm.com/)) --- [Troubleshooting custom
  domains](https://docs.github.com/articles/troubleshooting-custom-domains/#https-errors)

  HTTPS provides a layer of encryption that prevents others from
  snooping on or tampering with traffic to your site.
  When HTTPS is enforced, your site will only be served over HTTPS.
  [Learn more about securing your GitHub Pages site with
  HTTPS](https://docs.github.com/pages/getting-started-with-github-pages/securing-your-github-pages-site-with-https).

As discussed in https://github.com/git/git-scm.com/pull/1953, it would
appear to be outside our capabilities to fix this properly, therefore we
are unfortunately stuck with hard-coding a work-around: Manually
override the http:// URL produced by `actions/configure-pages` with the
equivalent https:// one.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>